Renames dicom classes
