+++ oct_converter/readers/__init__.py
@@ -1,8 +1,8 @@
 from .boct import BOCT
-from .dcm import Dicom
+from .dicom import Dicom
 from .e2e import E2E
 from .fda import FDA
 from .fds import FDS
 from .img import IMG
 from .poct import POCT
-from .zeissdcm import ZEISSDCM
+from .zeissdicom import ZEISSDicom
